Check required fields with validators for resource schemas
PushTopicSchema and StreamingChannelSchema should both accept
an input configuration where either a single field is specified
which uniquely identifies the resource or if multiple fields are
specified which can be used to construct the resource. To
validate such requirements validator methods are defined instead
of specifying the required status of fields.