Handle ConnectionError exceptions in RedisReplayStorage
Catch and log ConnectionError exceptions that might occur when
trying to get or set values in redis. This way, redis connection
issues won't affect the application, and it can continue to
process messages even if it can't store or retrieve replay markers.
Once the redis server is available again, the latest replay id
will be stored, when the next message arrives and the application
can recover as if the connection issue never happened.