Make ignorance of certain types of errors configurable
Make the ignorance of errors generated by external services
configurable. So users can choose between only logging the errors
and continuing with the operation or exiting from the application
as soon as the error occurs. These errors are network or internal
errors generated by the external services such as Salesforce,
redis and RabbitMQ.